reorganized components and styles
