+++ src/components/Profile/HomeContainer/styles.ts
@@ -1,7 +1,4 @@
-import React from 'react';
-import { Text, View, StyleSheet } from 'react-native';
-
-import HomeCard from './HomeCard';
+import { StyleSheet } from 'react-native';
 
 const styles = StyleSheet.create({
   mainContainer: {
@@ -54,24 +51,4 @@ const styles = StyleSheet.create({
   },
 });
 
-const HomeContainer = () => {
-  const badgeNumber = 10;
-
-  return (
-    <View style={styles.mainContainer}>
-      <View style={styles.mainLayout}>
-        <View style={styles.numberCircle}>
-          <Text style={styles.number}>{badgeNumber}</Text>
-        </View>
-        <Text style={styles.badgeText}>Total Badges Collected</Text>
-      </View>
-      <View style={styles.line} />
-      <Text style={styles.homeTitle}>HOME</Text>
-      <View style={styles.homesLayout}>
-        <HomeCard />
-      </View>
-    </View>
-  );
-};
-
-export default HomeContainer;
+export default styles;
