+++ src/components/Profile/SettingsHeader/index.tsx
@@ -1,30 +1,8 @@
 import React from 'react';
-import { Text, View, StyleSheet } from 'react-native';
+import { Text, View } from 'react-native';
 import { TouchableOpacity } from 'react-native-gesture-handler';
 import { useNavigation } from '@react-navigation/native';
-
-const styles = StyleSheet.create({
-  header: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    position: 'absolute',
-    alignItems: 'center',
-    width: '100%',
-    paddingTop: '10%',
-    paddingLeft: '5%',
-    paddingRight: '5%',
-    zIndex: 1,
-  },
-  profile: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    paddingLeft: 40,
-  },
-  done: {
-    fontSize: 15,
-    fontWeight: 'bold',
-  },
-});
+import styles from './styles';
 
 const SettingsHeader = () => {
   const navigation = useNavigation();
